API: add fullTextSearch query that works across all entities using EntityInfo (#5461)
