+++ demo/api/src/news/entities/news.entity.ts
@@ -4,12 +4,26 @@ import {
     CrudGenerator,
     DamImageBlock,
     EntityInfo,
+    entityToMikroOrmFullText,
     RootBlock,
     RootBlockDataScalar,
     RootBlockEntity,
     RootBlockType,
 } from "@comet/cms-api";
-import { BaseEntity, Collection, Embeddable, Embedded, Entity, Enum, OneToMany, OptionalProps, PrimaryKey, Property } from "@mikro-orm/postgresql";
+import {
+    BaseEntity,
+    Collection,
+    Embeddable,
+    Embedded,
+    Entity,
+    Enum,
+    FullTextType,
+    Index,
+    OneToMany,
+    OptionalProps,
+    PrimaryKey,
+    Property,
+} from "@mikro-orm/postgresql";
 import { Field, ID, InputType, ObjectType, registerEnumType } from "@nestjs/graphql";
 import { IsString } from "class-validator";
 import { v4 as uuid } from "uuid";
@@ -47,7 +61,7 @@ export class NewsContentScope {
     language: string;
 }
 
-@EntityInfo<News>({ name: "title", secondaryInformation: "slug", visible: { status: { $eq: NewsStatus.active } } })
+@EntityInfo<News>({ name: "title", secondaryInformation: "slug", visible: { status: { $eq: NewsStatus.active } }, fullText: "fullText" })
 @RootBlockEntity()
 @ObjectType()
 @Entity()
@@ -108,4 +122,12 @@ export class News extends BaseEntity {
     @Property({ onUpdate: () => new Date(), columnType: "timestamp with time zone" })
     @Field()
     updatedAt: Date = new Date();
+
+    @Index({ type: "fulltext" })
+    @Property<News>({
+        nullable: true,
+        type: new FullTextType(),
+        onUpdate: (news) => entityToMikroOrmFullText({ A: news.title, D: news.slug }, news.content),
+    })
+    fullText?: string;
 }
