+++ demo/api/src/products/entities/product.entity.ts
@@ -15,6 +15,8 @@ import {
     Collection,
     Entity,
     Enum,
+    FullTextType,
+    Index,
     ManyToMany,
     ManyToOne,
     OneToMany,
@@ -88,7 +90,12 @@ export class ProductPriceRange {
     max: number;
 }
 
-@EntityInfo<Product>({ name: "title", secondaryInformation: "manufacturer.name", visible: { status: { $eq: ProductStatus.Published } } })
+@EntityInfo<Product>({
+    name: "title",
+    secondaryInformation: "manufacturer.name",
+    visible: { status: { $eq: ProductStatus.Published } },
+    fullText: "fullText",
+})
 @ObjectType()
 @Entity()
 @RootBlockEntity<Product>({ isVisible: (product) => product.status === ProductStatus.Published })
@@ -242,4 +249,17 @@ export class Product extends BaseEntity implements ImportTargetInterface {
     @ManyToMany(() => DamFile)
     @Field(() => [DamFile])
     relatedImages = new Collection<DamFile>(this);
+
+    @Index({ type: "fulltext" })
+    @Property<Product>({
+        nullable: true,
+        type: new FullTextType(),
+        onUpdate: (page) => {
+            return {
+                A: page.title,
+                D: page.description,
+            };
+        },
+    })
+    fullText?: string;
 }
