+++ packages/api/cms-api/schema.gql
@@ -30,6 +30,7 @@ enum Permission {
   warnings
   sitePreview
   blockPreview
+  fullTextSearch
 }
 
 """
@@ -74,6 +75,8 @@ type CurrentUser {
 }
 
 type EntityInfo {
+  id: String!
+  entityName: String!
   name: String!
   secondaryInformation: String
 }
