+++ packages/api/cms-api/src/entity-info/entity-info.object.ts
@@ -8,9 +8,11 @@ import { Field, ObjectType } from "@nestjs/graphql";
 @ObjectType("EntityInfo")
 export class EntityInfoObject {
     @PrimaryKey({ type: "text" })
+    @Field()
     id: string;
 
     @PrimaryKey({ type: "text" })
+    @Field()
     entityName: string;
 
     @Field()
