+++ packages/api/cms-api/src/entity-info/entity-info.service.ts
@@ -1,9 +1,10 @@
-import { AnyEntity, EntityManager, EntityMetadata } from "@mikro-orm/postgresql";
+import { AnyEntity, EntityManager } from "@mikro-orm/postgresql";
 import { Injectable, Logger } from "@nestjs/common";
 
 import { DiscoverService } from "../dependencies/discover.service";
 import { ENTITY_INFO_METADATA_KEY, EntityInfo } from "./entity-info.decorator";
 import { EntityInfoObject } from "./entity-info.object";
+import { resolveFieldToSql } from "./resolve-field-to-sql";
 
 @Injectable()
 export class EntityInfoService {
@@ -14,111 +15,52 @@ export class EntityInfoService {
         private entityManager: EntityManager,
     ) {}
 
-    /**
-     * Resolves a field path (e.g., "title" or "manufacturer.name") to a SQL expression.
-     * Supports direct fields, embeddables, and n-level deep ManyToOne/OneToOne relations using subqueries.
-     */
-    private resolveFieldToSql(fieldPath: string, metadata: EntityMetadata, tableName: string): string {
-        const parts = fieldPath.split(".");
-
-        // Direct field - find the actual column name
-        if (parts.length === 1) {
-            const fieldProp = metadata.props.find((p) => p.name === parts[0]);
-            if (!fieldProp) {
-                throw new Error(`Field "${parts[0]}" not found in entity "${metadata.className}"`);
-            }
-
-            return `"${tableName}"."${fieldProp.fieldNames[0]}"`;
-        }
-
-        // Relation path - resolve first relation and recurse for the rest
-        const [relationName, ...remainingParts] = parts;
-
-        // Find the relation property in metadata
-        const relationProp = metadata.props.find((p) => p.name === relationName);
-
-        if (!relationProp) {
-            throw new Error(`Relation "${relationName}" not found in entity "${metadata.className}"`);
-        }
-
-        // Handle embedded properties - fields are columns on the same table
-        if (relationProp.kind === "embedded") {
-            let currentProp = relationProp;
-            let currentName = relationName;
-            for (const part of remainingParts) {
-                const childProp = currentProp.embeddedProps?.[part];
-                if (!childProp) {
-                    throw new Error(`Embedded field "${part}" not found in embeddable "${currentName}" of entity "${metadata.className}"`);
-                }
-                currentName = part;
-                currentProp = childProp;
-            }
-            return `"${tableName}"."${currentProp.fieldNames[0]}"`;
-        }
-
-        if (relationProp.kind !== "m:1" && relationProp.kind !== "1:1") {
-            throw new Error(
-                `Only ManyToOne, OneToOne relations and embeddables are supported for EntityInfo. "${relationName}" is "${relationProp.kind}"`,
-            );
-        }
-
-        if (!relationProp.targetMeta) {
-            throw new Error(`Relation "${relationName}" has no target metadata`);
-        }
-
-        // Get the join column (foreign key) and target table info
-        const joinColumn = relationProp.joinColumns[0];
-        const targetTableName = relationProp.targetMeta.tableName;
-        const targetPrimaryKey = relationProp.targetMeta.primaryKeys[0];
-
-        // Recursively resolve the remaining path in the target entity
-        const innerSql = this.resolveFieldToSql(remainingParts.join("."), relationProp.targetMeta, targetTableName);
-
-        return `(SELECT ${innerSql} FROM "${targetTableName}" WHERE "${targetTableName}"."${targetPrimaryKey}" = "${tableName}"."${joinColumn}")`;
-    }
-
     async createEntityInfoView(): Promise<void> {
         const indexSelects: string[] = [];
         const targetEntities = this.discoverService.discoverTargetEntities();
         for (const targetEntity of targetEntities) {
             const entityInfo = Reflect.getMetadata(ENTITY_INFO_METADATA_KEY, targetEntity.entity) as EntityInfo<AnyEntity>;
-            if (entityInfo) {
-                if (typeof entityInfo === "string") {
-                    indexSelects.push(entityInfo);
-                } else {
-                    const { entityName, metadata } = targetEntity;
-                    const primary = metadata.primaryKeys[0];
+            if (!entityInfo) {
+                continue;
+            }
 
-                    // Resolve the name field (must not be NULL)
-                    const nameSql = `COALESCE(${this.resolveFieldToSql(entityInfo.name, metadata, metadata.tableName)}, '')`;
+            if (typeof entityInfo === "string") {
+                indexSelects.push(
+                    `SELECT sub."name", sub."secondaryInformation", sub."visible", sub."id", sub."entityName" FROM (${entityInfo}) sub`,
+                );
+            } else {
+                const { entityName, metadata } = targetEntity;
+                const primary = metadata.primaryKeys[0];
 
-                    // Resolve the secondaryInformation field (if provided, can be NULL)
-                    let secondaryInformationSql = "null";
-                    if (entityInfo.secondaryInformation) {
-                        secondaryInformationSql = this.resolveFieldToSql(entityInfo.secondaryInformation, metadata, metadata.tableName);
+                // Resolve the name field (must not be NULL)
+                const nameSql = `COALESCE(${resolveFieldToSql(entityInfo.name, metadata, metadata.tableName)}, '')`;
+
+                // Resolve the secondaryInformation field (if provided, can be NULL)
+                let secondaryInformationSql = "null";
+                if (entityInfo.secondaryInformation) {
+                    secondaryInformationSql = resolveFieldToSql(entityInfo.secondaryInformation, metadata, metadata.tableName);
+                }
+
+                let visibleSql = "true";
+                if (entityInfo.visible) {
+                    const qb = this.entityManager.createQueryBuilder(targetEntity.entity.name, metadata.tableName);
+                    const query = qb.select("*").where(entityInfo.visible);
+                    const sql = query.getFormattedQuery();
+                    const sqlWhereMatch = sql.match(/^select .*? from .*? where (.*)/);
+                    if (!sqlWhereMatch) {
+                        throw new Error(`Could not extract where clause from query: ${sql}`);
                     }
+                    visibleSql = sqlWhereMatch[1];
+                }
 
-                    let visibleSql = "true";
-                    if (entityInfo.visible) {
-                        const qb = this.entityManager.createQueryBuilder(targetEntity.entity.name, metadata.tableName);
-                        const query = qb.select("*").where(entityInfo.visible);
-                        const sql = query.getFormattedQuery();
-                        const sqlWhereMatch = sql.match(/^select .*? from .*? where (.*)/);
-                        if (!sqlWhereMatch) {
-                            throw new Error(`Could not extract where clause from query: ${sql}`);
-                        }
-                        visibleSql = sqlWhereMatch[1];
-                    }
-
-                    const select = `SELECT
+                const select = `SELECT
                                 ${nameSql} "name",
                                 ${secondaryInformationSql} "secondaryInformation",
                                 ${visibleSql} AS "visible",
                                 "${metadata.tableName}"."${primary}"::text "id",
                                 '${entityName}' "entityName"
                             FROM "${metadata.tableName}"`;
-                    indexSelects.push(select);
-                }
+                indexSelects.push(select);
             }
         }
